fix - refactor key endpoints
