+++ litellm/tests/test_blocked_user_list.py
@@ -29,18 +29,19 @@ import pytest, logging, asyncio
 import litellm, asyncio
 from litellm.proxy.proxy_server import (
     new_user,
-    generate_key_fn,
     user_api_key_auth,
     user_update,
+    user_info,
+    block_user,
+)
+from litellm.proxy.management_endpoints.key_management_endpoints import (
     delete_key_fn,
     info_key_fn,
     update_key_fn,
     generate_key_fn,
     generate_key_helper_fn,
-    user_info,
-    block_user,
 )
-from litellm.proxy.spend_management_endpoints import (
+from litellm.proxy.spend_reporting_endpoints.spend_management_endpoints import (
     spend_user_fn,
     spend_key_fn,
     view_spend_logs,
